tungnv_b23dckh130_training_gd1/Week 4/BE/Toblug/requirements.txt
fastapi-mail
pydantic-settings
redis
markdown
python-slugify
pydantic
python-jose>=3.5.0
passlib>=1.7.4
fastapi>=0.120.2
sqlalchemy>=2.0.44
psycopg2>=2.9.11
python-dotenv>=1.0.0
uvicorn>=0.30.0
jose>=1.0.0
alembic>=1.13.0
celery>=5.3.0
nest-asyncio>=1.5.8